Update the project as I change my computer.

+++ src/main/java/com/skillmentor/root/repository/MentorRepository.java
@@ -2,6 +2,11 @@ package com.skillmentor.root.repository;
 
 import com.skillmentor.root.entity.MentorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface MentorRepository extends JpaRepository<MentorEntity, Long> {
+import java.util.List;
+import java.util.Optional;
+
+public interface MentorRepository extends JpaRepository<MentorEntity, Integer> {
 }
